Created new pillow block for X rotary bearing
A new pillow block, similar to the one in the Y axis, replaced the old version.
The Y rotary bearing block has been modified with a thinner base to allow for M5 10mm screws. It also has a chamfer for support.
The secondary X axis has been hidden in the assembly for now.